+++ packages/terra-framework-docs/src/terra-dev-site/test/date-time-picker/DateTimePickerDefaultDSTTwelveHour.test.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import DateTimeUtil from 'terra-date-time-picker/lib/DateTimeUtils';
+import DateTimePickerExampleTemplate from './common/DateTimePickerExampleTemplate';
+
+const DateTimePickerDefault = () => (
+  <DateTimePickerExampleTemplate
+    initialTimeZone="America/Chicago"
+    timeVariant={DateTimeUtil.FORMAT_12_HOUR}
+  />
+);
+
+export default DateTimePickerDefault;