+++ packages/terra-time-input/tests/wdio/time-input-spec.js
@@ -1,76 +1,79 @@
+const validateElement = (testName) => Terra.validates.element(testName, { selector: '#timeInput' });
+const validateRoot = (testName) => Terra.validates.element(testName);
+
 Terra.describeViewports('Time Input', ['medium'], () => {
   it('works with a label', () => {
     browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/accessibility');
-    Terra.validates.element('default no time');
+    validateElement('default with a label');
   });
 
   it('displays default with no time provided', () => {
     browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/default');
-    Terra.validates.element('default no time');
+    validateElement('default no time');
   });
 
   it('displays default with time provided', () => {
     browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/default-time');
-    Terra.validates.element('default with time');
+    validateElement('default with time');
   });
 
   it('displays second with no time provided', () => {
     browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/second');
-    Terra.validates.element('seconds no time');
+    validateElement('seconds no time');
   });
 
   it('displays second with time provided', () => {
     browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/second-time');
-    Terra.validates.element('second with time');
+    validateElement('second with time');
   });
 
   it('displays invalid time provided', () => {
     browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/invalid-provided-time');
-    Terra.validates.element('invalid time');
+    validateElement('invalid time');
   });
 
   it('displays invalid time input', () => {
     browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/invalid-time');
-    Terra.validates.element('invalid input');
+    validateElement('invalid input');
   });
 
   it('displays invalid mobile time input', () => {
     browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/mobile-invalid');
-    Terra.validates.element('invalid mobile input');
+    validateElement('invalid mobile input');
   });
 
   it('displays incomplete time input', () => {
     browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/incomplete-time');
-    Terra.validates.element('incomplete input');
+    validateElement('incomplete input');
   });
 
   it('displays incomplete mobile time input', () => {
     browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/mobile-incomplete');
-    Terra.validates.element('incomplete mobile');
+    validateElement('incomplete mobile');
   });
 
   it('displays invalid and incomplete time input', () => {
     browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/invalid-incomplete-time');
-    Terra.validates.element('invalid and incomplete');
+    validateElement('invalid and incomplete');
   });
 
   it('displays invalid and incomplete mobile time input', () => {
     browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/mobile-invalid-incomplete');
-    Terra.validates.element('invalid and incomplete mobile');
+    validateElement('invalid and incomplete mobile');
   });
 
   it('displays focus hour styles', () => {
     browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/default-time');
     $('#timeInput input[name="terra-time-hour-time-input"]').click();
     Terra.hideInputCaret('#timeInput input[name="terra-time-hour-time-input"]');
-    Terra.validates.element('focus hour styles');
+    validateElement('focus hour styles');
   });
 
   it('displays focus time component styles', () => {
     browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/incomplete-time');
     $('#timeInput input[name="terra-time-hour-time-input"]').click();
     Terra.hideInputCaret('#timeInput input[name="terra-time-hour-time-input"]');
-    Terra.validates.element('focus time component styles');
+    validateElement('focus time component styles');
   });
 
   it('displays focus minute styles', () => {
@@ -78,7 +81,7 @@ Terra.describeViewports('Time Input', ['medium'], () => {
     browser.refresh();
     $('#timeInput input[name="terra-time-minute-time-input"]').click();
     Terra.hideInputCaret('#timeInput input[name=\'terra-time-minute-time-input\']');
-    Terra.validates.element('focus minute styles');
+    validateElement('focus minute styles');
   });
 
   it('displays focus second styles', () => {
@@ -87,7 +90,7 @@ Terra.describeViewports('Time Input', ['medium'], () => {
     $('#timeInput input[name="terra-time-second-time-input"]').click();
     Terra.hideInputCaret('#timeInput input[name="terra-time-second-time-input"]');
 
-    Terra.validates.element('focus second styles');
+    validateElement('focus second styles');
   });
 });
 
@@ -98,7 +101,7 @@ Terra.describeViewports('Time input onChange operations', ['medium'], () => {
 
     $('#timeInput input[name="terra-time-hour-time-input"]').click();
     browser.keys('123456');
-    Terra.validates.element('trigger onChange');
+    validateRoot('trigger onChange');
   });
 
   it('prepends 0 on single digit hour', () => {
@@ -109,7 +112,7 @@ Terra.describeViewports('Time input onChange operations', ['medium'], () => {
     $('#timeInput input[name="terra-time-hour-time-input"]').click();
     browser.keys('2');
     $('#timeInput input[name="terra-time-minute-time-input"]').click();
-    Terra.validates.element('prepend 0 to single digit');
+    validateRoot('prepend 0 to single digit');
   });
 
   it('prepends 0 on hour input of 3 or more', () => {
@@ -119,7 +122,7 @@ Terra.describeViewports('Time input onChange operations', ['medium'], () => {
 
     $('#timeInput input[name="terra-time-hour-time-input"]').click();
     browser.keys('3');
-    Terra.validates.element('prepend 0 to hour more than 3');
+    validateRoot('prepend 0 to hour more than 3');
   });
 
   it('displays time input does not prepend 0 on hour input less than 3', () => {
@@ -129,7 +132,7 @@ Terra.describeViewports('Time input onChange operations', ['medium'], () => {
 
     $('#timeInput input[name="terra-time-hour-time-input"]').click();
     browser.keys('2');
-    Terra.validates.element('does not prepend 0 to hour less than 3');
+    validateRoot('does not prepend 0 to hour less than 3');
   });
 
   it('displays time input does not accept hour input greater than 23', () => {
@@ -139,7 +142,7 @@ Terra.describeViewports('Time input onChange operations', ['medium'], () => {
 
     $('#timeInput input[name="terra-time-hour-time-input"]').click();
     browser.keys('24');
-    Terra.validates.element('does not accept hour more than 23');
+    validateRoot('does not accept hour more than 23');
   });
 
   it('prepends 0 on single digit minute', () => {
@@ -150,7 +153,7 @@ Terra.describeViewports('Time input onChange operations', ['medium'], () => {
     $('#timeInput input[name="terra-time-minute-time-input"]').click();
     browser.keys('2');
     $('#timeInput input[name="terra-time-hour-time-input"]').click();
-    Terra.validates.element('prepends 0 on single digit minute');
+    validateRoot('prepends 0 on single digit minute');
   });
 
   it('prepends 0 on minute input of 6 or more', () => {
@@ -160,7 +163,7 @@ Terra.describeViewports('Time input onChange operations', ['medium'], () => {
 
     $('#timeInput input[name="terra-time-minute-time-input"]').click();
     browser.keys('6');
-    Terra.validates.element('prepends 0 on minute input of 6 or more');
+    validateRoot('prepends 0 on minute input of 6 or more');
   });
 
   it('does not prepend 0 on minute input less than 6', () => {
@@ -170,7 +173,7 @@ Terra.describeViewports('Time input onChange operations', ['medium'], () => {
 
     $('#timeInput input[name="terra-time-minute-time-input"]').click();
     browser.keys('5');
-    Terra.validates.element('does not prepend 0 to minute');
+    validateRoot('does not prepend 0 to minute');
   });
 
   it('does not accept minute input greater than 59', () => {
@@ -180,7 +183,7 @@ Terra.describeViewports('Time input onChange operations', ['medium'], () => {
 
     $('#timeInput input[name="terra-time-minute-time-input"]').click();
     browser.keys('66');
-    Terra.validates.element('does not accept minute input greater than 59');
+    validateRoot('does not accept minute input greater than 59');
   });
 
   it('Pressing DELETE in minute with no value focuses to hour', () => {
@@ -191,7 +194,7 @@ Terra.describeViewports('Time input onChange operations', ['medium'], () => {
 
     $('#timeInput input[name="terra-time-minute-time-input"]').click();
     browser.keys(['Delete']);
-    Terra.validates.element('delete focuses hour');
+    validateRoot('delete focuses hour');
   });
 
   it('prepends 0 on single digit second', () => {
@@ -202,7 +205,7 @@ Terra.describeViewports('Time input onChange operations', ['medium'], () => {
     $('#timeInput input[name="terra-time-second-time-input"]').click();
     browser.keys('2');
     $('#timeInput input[name="terra-time-hour-time-input"]').click();
-    Terra.validates.element('prepends 0 on single digit second');
+    validateRoot('prepends 0 on single digit second');
   });
 
   it('prepends 0 on second input of 6 or more', () => {
@@ -212,7 +215,7 @@ Terra.describeViewports('Time input onChange operations', ['medium'], () => {
 
     $('#timeInput input[name="terra-time-second-time-input"]').click();
     browser.keys('6');
-    Terra.validates.element('prepends 0 on second input of 6 or more');
+    validateRoot('prepends 0 on second input of 6 or more');
   });
 
   it('does not prepend 0 on second input less than 6', () => {
@@ -222,7 +225,7 @@ Terra.describeViewports('Time input onChange operations', ['medium'], () => {
 
     $('#timeInput input[name="terra-time-second-time-input"]').click();
     browser.keys('5');
-    Terra.validates.element('does not prepend 0 on second input less than 6');
+    validateRoot('does not prepend 0 on second input less than 6');
   });
 
   it('does not accept second input greater than 59', () => {
@@ -232,7 +235,7 @@ Terra.describeViewports('Time input onChange operations', ['medium'], () => {
 
     $('#timeInput input[name="terra-time-second-time-input"]').click();
     browser.keys('66');
-    Terra.validates.element('does not accept second input greater than 59');
+    validateRoot('does not accept second input greater than 59');
   });
 
   it('Pressing DELETE in second with no value focuses to minute', () => {
@@ -243,7 +246,7 @@ Terra.describeViewports('Time input onChange operations', ['medium'], () => {
     $('#timeInput input[name="terra-time-second-time-input"]').click();
     browser.keys(['Delete']);
 
-    Terra.validates.element('delete focuses minute');
+    validateRoot('delete focuses minute');
   });
 
   it('Triggers an onChange for onBlur on the minute input', () => {
@@ -253,7 +256,7 @@ Terra.describeViewports('Time input onChange operations', ['medium'], () => {
 
     $('#timeInput input[name="terra-time-hour-time-input"]').click();
     browser.keys('747');
-    Terra.validates.element('trigger onChange for onBlur');
+    validateRoot('trigger onChange for onBlur');
   });
 
   it('Does not trigger onChange for an hour with just one digit', () => {
@@ -265,7 +268,7 @@ Terra.describeViewports('Time input onChange operations', ['medium'], () => {
     browser.keys('1223');
     $('#timeInput input[name="terra-time-hour-time-input"]').click();
     browser.keys(['ArrowLeft', 'ArrowRight', 'Delete']);
-    Terra.validates.element('onChange not triggered for hour with one digit');
+    validateRoot('onChange not triggered for hour with one digit');
   });
 
   it('Does not trigger onChange for a minute with just one digit', () => {
@@ -277,7 +280,7 @@ Terra.describeViewports('Time input onChange operations', ['medium'], () => {
     browser.keys('1223');
     $('#timeInput input[name="terra-time-minute-time-input"]').click();
     browser.keys(['ArrowLeft', 'ArrowRight', 'Delete']);
-    Terra.validates.element('onChange not triggered for minute with one digit');
+    validateRoot('onChange not triggered for minute with one digit');
   });
 });
 
@@ -321,7 +324,7 @@ describe('Time Input shortcut key operations', () => {
     browser.keys('23');
     $('#timeInput input[name="terra-time-hour-time-input"]').click();
     browser.keys(['ArrowDown']);
-    Terra.validates.element('DOWN_ARROW decrements hour by 1');
+    validateRoot('DOWN_ARROW decrements hour by 1');
   });
 
   it('ignores DOWN_ARROW when the hour has reached 00', () => {
@@ -333,7 +336,7 @@ describe('Time Input shortcut key operations', () => {
     browser.keys('00');
     $('#timeInput input[name="terra-time-hour-time-input"]').click();
     browser.keys(['ArrowDown']);
-    Terra.validates.element('DOWN_ARROW ignored for hour 00');
+    validateRoot('DOWN_ARROW ignored for hour 00');
   });
 
   it('increments hour by 1 using UP_ARROW ', () => {
@@ -343,7 +346,7 @@ describe('Time Input shortcut key operations', () => {
 
     $('#timeInput input[name="terra-time-hour-time-input"]').click();
     browser.keys(['ArrowUp', 'ArrowUp']);
-    Terra.validates.element('UP_ARROW increments hour by 1');
+    validateRoot('UP_ARROW increments hour by 1');
   });
 
   it('ignores UP_ARROW when the hour has reached 23', () => {
@@ -355,10 +358,10 @@ describe('Time Input shortcut key operations', () => {
     browser.keys('23');
     $('#timeInput input[name="terra-time-hour-time-input"]').click();
     browser.keys(['ArrowUp']);
-    Terra.validates.element('UP_ARROW ignored for hour 23');
+    validateRoot('UP_ARROW ignored for hour 23');
   });
 
-  describe.only('pressing N', () => {
+  describe('pressing N', () => {
     describe('in the hour input', () => {
       it('should set time to current if blank', () => {
         browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/default');
@@ -516,7 +519,7 @@ describe('Time Input shortcut key operations', () => {
     });
   });
 
-  describe.only('pressing -', () => {
+  describe('pressing -', () => {
     describe('in the hour input', () => {
       it('should set time to current time minus one minute if blank', () => {
         browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/default');
@@ -606,6 +609,21 @@ describe('Time Input shortcut key operations', () => {
         expect($('#timeInput input[name="terra-time-minute-time-input"]')).toHaveValue('59');
       });
 
+      it('should roll over to 23 if hour and minute is at 00:00', () => {
+        browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/default');
+        browser.refresh();
+        Terra.hideInputCaret('#timeInput input[name="terra-time-minute-time-input"]');
+
+        $('#timeInput input[name="terra-time-hour-time-input"]').click();
+        browser.keys('00');
+        $('#timeInput input[name="terra-time-minute-time-input"]').click();
+        browser.keys('00');
+        $('#timeInput input[name="terra-time-minute-time-input"]').click();
+        browser.keys('-');
+        expect($('#timeInput input[name="terra-time-hour-time-input"]')).toHaveValue('23');
+        expect($('#timeInput input[name="terra-time-minute-time-input"]')).toHaveValue('59');
+      });
+
       it('should subtract 1 minute if filled', () => {
         browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/default-time');
         browser.refresh();
@@ -671,7 +689,7 @@ describe('Time Input shortcut key operations', () => {
     });
   });
 
-  describe.only('pressing +', () => {
+  describe('pressing +', () => {
     describe('in the hour input', () => {
       it('should set time to current time plus one minute if blank', () => {
         browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/default');
@@ -758,6 +776,21 @@ describe('Time Input shortcut key operations', () => {
         $('#timeInput input[name="terra-time-minute-time-input"]').click();
         browser.keys('+');
         expect($('#timeInput input[name="terra-time-hour-time-input"]')).toHaveValue('12');
+        expect($('#timeInput input[name="terra-time-minute-time-input"]')).toHaveValue('00');
+      });
+
+      it('should roll over the hour to 00 if hour and minute is at 23:59', () => {
+        browser.url('/raw/tests/cerner-terra-framework-docs/time-input/time-input/default');
+        browser.refresh();
+        Terra.hideInputCaret('#timeInput input[name="terra-time-minute-time-input"]');
+
+        $('#timeInput input[name="terra-time-hour-time-input"]').click();
+        browser.keys('23');
+        $('#timeInput input[name="terra-time-minute-time-input"]').click();
+        browser.keys('59');
+        $('#timeInput input[name="terra-time-minute-time-input"]').click();
+        browser.keys('+');
+        expect($('#timeInput input[name="terra-time-hour-time-input"]')).toHaveValue('00');
         expect($('#timeInput input[name="terra-time-minute-time-input"]')).toHaveValue('00');
       });