+++ packages/terra-dialog-modal/CHANGELOG.md
@@ -5,6 +5,11 @@
 * Changed
   * Additional terra-dialog-modal deprecation notices
 
+## 3.74.4 - (March 29, 2022)
+
+* Changed
+  * Minor dependency version bump
+
 ## 3.74.3 - (March 15, 2022)
 
 * Changed